restructuring components for routing
