+++ src/App.css
@@ -3,37 +3,6 @@ img{
 	margin-right: auto;
 }
 
-.Contain{
-	display: flex;
-	flex-direction: column;
-	justify-content: baseline;
-	align-items: center;
-	width: 86vw;
-	max-width: 840px;
-	border-radius: 16px 16px;
-	margin-top: 1.4%;
-	padding: 1% 3%;
-	min-height: 87vh;
-}
-
-.Canvas{
-	margin-top: 3.9rem;
-	display: flex;
-	justify-content: center;
-	padding-bottom: 1.5%;
-}
-
-@media (max-width: 640px) { 
-	.Contain{
-		border-radius: 0;
-		padding: 1% 3%;
-		width: 100vw;
-		min-height: 95vh;
-	}
-	.Canvas{
-		padding-bottom: 0;
-	}
-}
 
 .App-up{
 	background-color: hsl(var(--a));
