+++ src/Nav.js
@@ -1,21 +1,18 @@
 import "./Nav.css";
 import Category from "./Category";
-// import smallLogo from "./assets/smallLogo.png"
-// import bigLogo from "./assets/bigLogo.png"
+import Logo from "./assets/logo49.png"
+// import React, { useState } from 'react';
+
 function Nav({reqType, setReqType}) {
+    // const [searchState, setSearchState] = useState(false);
 
     return ( 
         <div className="navbar bg-base-200 text-primary-content">
-            <div className="flex-1">
-                <a href='https://news.ycombinator.com/' className="btn Bigbtn no-animation btn-outline btn-accent text-xl Marginleft">HackerNews</a>
-                <a href='https://news.ycombinator.com/' className="btn Smallbtn no-animation btn-outline btn-accent text-xl Marginleft">HN</a>
-                {/* <a href="" className="Bigbtn Marginleft"> <img src={bigLogo}></img> </a>
-                <a href="" className="Smallbtn Marginleft"> <img src={smallLogo}></img> </a> */}
-            </div>
+            <a href='/' className="Logo"> <img src={Logo} alt="logo"></img> </a>
 
             <form 
             onSubmit={(e) => e.preventDefault()} 
-            className="join Marginright"
+            className="join btnGroup"
             >
                 <Category
                 buttonText="top"
@@ -34,13 +31,11 @@ function Nav({reqType, setReqType}) {
                 />
             </form>
 
-            
-
-            {/* <div className="flex-none Marginright">
-                <div className="form-control">
-                <input type="text" placeholder="Search" className="input input-accent Searchbar" />
+            {/* <button onClick={() => {setSearchState(true);}} className="btn btn-accent" style={{display: searchState ? 'none' : 'block'}}>&#x1F50D;</button>
+            <div className="">
+                <div className="form-control" style={{display: searchState ? 'block' : 'none'}}>
+                    <input type="text" placeholder="&#x1F50D;" className="input input-accent Searchbar" style={{textAlign: "right"}}/>
                 </div>
-                
             </div> */}
         </div>
     );
