+++ src/pages/Comments.js
@@ -1,6 +1,10 @@
 function Comments() {
+    // console.log(window.location.pathname)
     return ( 
-        <h1> Welcome to the comments!!! </h1>
+        <div style={{position: "absolute", top: "40%", left: "40%"}}>
+            <h1> Welcome to the comments!!! </h1>
+            <h2>The comment page is yet to be developed...</h2>
+        </div>
     );
 }
