+++ src/pages/Content.css
@@ -1,6 +1,6 @@
 
 .card-title{
-    word-break: break-all;
+    word-break: break-word;
 }
 
 .New-card{
